pyproject.toml
[project]
name = "aqueduct"
requires-python = ">=3.10,<3.14"
version = "0.1.0"
dependencies = [
    "google-photos-library-api>=0.9.0",
    "graphviz>=0.21",
    "instaloader>=4.10",
    "notion-client>=2.2.1",
    "pdbpp>=0.11.7",
    "prefect[github]>=3.5.0",
    "python-dotenv>=1.2.1",
    "requests>=2.31.0",
    "synology-api>=0.6.0",
    "tweepy>=4.14.0",
]


[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
force-include = { "pyproject.toml" = "pyproject.toml" }

[tool.dg]
directory_type = "project"

[tool.dg.project]
root_module = "aqueduct"
registry_modules = [
    "aqueduct.components.*",
]


